python-backend/requirements.txt
transformers
torch
tokenizers
fastapi
uvicorn
python-multipart
pypdf
langchain
numpy
sentence_transformers
sentencepiece
accelerate
chromadb
tiktoken
